+++ o_el_iks.API/ITokenProvider.cs
@@ -1,0 +1,6 @@
+﻿namespace o_el_iks.API;
+
+public interface ITokenProvider
+{
+    string GenerateToken();
+}
