+++ o_el_iks.API/TokenProvider.cs
@@ -1,0 +1,9 @@
+﻿namespace o_el_iks.API;
+
+public class TokenProvider : ITokenProvider
+{
+    public string GenerateToken()
+    {
+        return "zalogowany";
+    }
+}
